split ssz static and generic tests
